+++ src/main/java/fr/tropweb/miningmanager/Utils.java
@@ -1,6 +1,7 @@
 package fr.tropweb.miningmanager;
 
 import fr.tropweb.miningmanager.commands.struct.CommandManager;
+import fr.tropweb.miningmanager.commands.struct.PermissionManager;
 import fr.tropweb.miningmanager.commands.struct.SubCommand;
 import org.bukkit.ChatColor;
 import org.bukkit.entity.Player;
@@ -13,26 +14,30 @@ import java.nio.charset.StandardCharsets;
 
 public class Utils {
 
-    public static void green(Player player, String message, Object... args) {
+    public static void green(final Player player, final String message, final Object... args) {
         player.sendMessage(String.format(ChatColor.GREEN + message, args));
     }
 
-    public static void red(Player player, String message, Object... args) {
+    public static void red(final Player player, final String message, final Object... args) {
         player.sendMessage(String.format(ChatColor.RED + message, args));
     }
 
-    public static boolean hasPerm(Player player, SubCommand subCommand) {
-        return hasPerm(player, subCommand.permission());
+    public static boolean hasPerm(final Player player, final SubCommand subCommand) {
+        return hasPerm(player, subCommand.getPermissionManager());
+    }
+
+    public static boolean hasPerm(final Player player, final CommandManager commandManager) {
+        return hasPerm(player, commandManager.getPermissionManager());
     }
 
     /**
      * Get if the player have the permission.
      *
-     * @param player     player bukkit object
-     * @param permission enum of permission
+     * @param player            player bukkit object
+     * @param permissionManager enum of permission
      * @return true if the player have the permission or if he is OP or return false
      */
-    public static boolean hasPerm(final Player player, final CommandManager permission) {
+    public static boolean hasPerm(final Player player, final PermissionManager permissionManager) {
         // if player
         if (player == null) return false;
 
@@ -40,7 +45,7 @@ public class Utils {
         if (player.isOp()) return true;
 
         // Check permission
-        return player.hasPermission(permission.getPermission());
+        return player.hasPermission(permissionManager.getPermission());
     }
 
     /**