+++ src/main/java/fr/tropweb/miningmanager/dao/impl/AbstractDAO.java
@@ -1,6 +1,6 @@
 package fr.tropweb.miningmanager.dao.impl;
 
-import fr.tropweb.miningmanager.dao.fields.BlockLiteAbstractFields;
+import fr.tropweb.miningmanager.dao.fields.BlockDataFields;
 import fr.tropweb.miningmanager.exception.DatabaseException;
 import org.apache.commons.lang.StringEscapeUtils;
 import org.bukkit.Material;
@@ -12,21 +12,21 @@ import java.util.StringJoiner;
 
 public class AbstractDAO {
 
-    protected static int getInt(final ResultSet resultSet, final BlockLiteAbstractFields blockLiteFields) throws SQLException {
-        return resultSet.getInt(blockLiteFields.getFieldName());
+    protected static int getInt(final ResultSet resultSet, final BlockDataFields blockDataFields) throws SQLException {
+        return resultSet.getInt(blockDataFields.getFieldName());
     }
 
-    protected static String getString(final ResultSet resultSet, final BlockLiteAbstractFields blockLiteFields) throws SQLException {
-        return resultSet.getString(blockLiteFields.getFieldName());
+    protected static String getString(final ResultSet resultSet, final BlockDataFields blockDataFields) throws SQLException {
+        return resultSet.getString(blockDataFields.getFieldName());
     }
 
-    protected static boolean getBoolean(final ResultSet resultSet, final BlockLiteAbstractFields blockLiteFields) throws SQLException {
-        return resultSet.getBoolean(blockLiteFields.getFieldName());
+    protected static boolean getBoolean(final ResultSet resultSet, final BlockDataFields blockDataFields) throws SQLException {
+        return resultSet.getBoolean(blockDataFields.getFieldName());
     }
 
-    protected static Material getMaterial(final ResultSet resultSet, final BlockLiteAbstractFields blockLiteFields) throws SQLException {
+    protected static Material getMaterial(final ResultSet resultSet, final BlockDataFields blockDataFields) throws SQLException {
         try {
-            return Material.valueOf(resultSet.getString(blockLiteFields.getFieldName()));
+            return Material.valueOf(resultSet.getString(blockDataFields.getFieldName()));
         } catch (final IllegalArgumentException e) {
             throw new DatabaseException(e);
         }