+++ src/main/java/fr/tropweb/miningmanager/engine/SQLiteEngine.java
@@ -1,21 +1,21 @@
 package fr.tropweb.miningmanager.engine;
 
 import fr.tropweb.miningmanager.dao.BlockDAO;
-import fr.tropweb.miningmanager.dao.impl.BlockLiteDAO;
+import fr.tropweb.miningmanager.dao.impl.BlockDataDAO;
 import fr.tropweb.miningmanager.dao.sqlite.SQLiteDAO;
-import fr.tropweb.miningmanager.pojo.BlockLite;
+import fr.tropweb.miningmanager.pojo.BlockData;
 
 /**
  * This class manage the object <code>SQLiteDAO</code> and provide all DAO we need to use the database.
  */
 public final class SQLiteEngine {
-    private final BlockDAO<BlockLite> blockDAO;
+    private final BlockDAO<BlockData> blockDAO;
 
     public SQLiteEngine(final SQLiteDAO sqliteDAO) {
-        this.blockDAO = new BlockLiteDAO(sqliteDAO.getConnection());
+        this.blockDAO = new BlockDataDAO(sqliteDAO.getConnection());
     }
 
-    public BlockDAO<BlockLite> getBlockDAO() {
+    public BlockDAO<BlockData> getBlockDAO() {
         return this.blockDAO;
     }
 }