+++ CHANGELOG.md
@@ -1,15 +1,30 @@
 # ChangeLog
 
-**implemented:**
+## [Release 1.1.0](https://github.com/Genskao/MiningManager/releases/tag/v1.1.0)
 
-- Economy plugins (VaultAPI)
+**Refactoring:**
+
+- Rename enum and class
+
+**Implemented:**
+
+- Economy plugin (VaultAPI)
 - Prices for scan and mining
 - Upgrade permissions
+- Towny plugin
+- Permission to ignore price and towny
 
 **Fixed bugs:**
 
 - Start mining task scan the chunk before the mining
 - Mining not remove block regenerated or added while the mining task
+- Auto scan not show the price of the scan
+- Block placed broken not remove from db
+
+**Merged pull requests:**
+
+- [Add economy plugin (VaultAPI)](https://github.com/Genskao/MiningManager/pull/9)
+- [Add towny plugin](https://github.com/Genskao/MiningManager/pull/10)
 
 ## [Release 1.0.6](https://github.com/Genskao/MiningManager/releases/tag/v1.0.6)